Rename everything to spacectl
