+++ .gitignore
@@ -4,7 +4,7 @@
 *.dll
 *.so
 *.dylib
-spacelift-cli
+spacectl
 
 # Test binary, built with `go test -c`
 *.test
