+++ internal/cmd/authenticated/client.go
@@ -3,8 +3,8 @@ package authenticated
 import (
 	"github.com/urfave/cli/v2"
 
-	"github.com/spacelift-io/spacelift-cli/client"
-	"github.com/spacelift-io/spacelift-cli/client/session"
+	"github.com/spacelift-io/spacectl/client"
+	"github.com/spacelift-io/spacectl/client/session"
 )
 
 // Client is the authenticated client that can be used by all CLI commands.
