+++ internal/cmd/profile/login_command.go
@@ -12,7 +12,7 @@ import (
 	"github.com/urfave/cli/v2"
 	"golang.org/x/term"
 
-	"github.com/spacelift-io/spacelift-cli/client/session"
+	"github.com/spacelift-io/spacectl/client/session"
 )
 
 func loginCommand() *cli.Command {
