+++ internal/cmd/profile/logout_command.go
