+++ internal/cmd/stack/run_trigger.go
@@ -7,8 +7,8 @@ import (
 	"github.com/shurcooL/graphql"
 	"github.com/urfave/cli/v2"
 
-	"github.com/spacelift-io/spacelift-cli/client/structs"
-	"github.com/spacelift-io/spacelift-cli/cmd/internal/authenticated"
+	"github.com/spacelift-io/spacectl/client/structs"
+	"github.com/spacelift-io/spacectl/internal/cmd/authenticated"
 )
 
 func runTrigger(spaceliftType, humanType string) cli.ActionFunc {
